requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
build==0.10.0
    # via pip-tools
click==8.1.4
    # via pip-tools
colorama==0.4.6
    # via
    #   -r requirements.in
    #   build
    #   click
    #   pytest
exceptiongroup==1.1.2
    # via pytest
iniconfig==2.0.0
    # via pytest
jsonschema==4.18.2
    # via -r requirements.in
jsonschema-specifications==2023.6.1
    # via jsonschema
markupsafe==2.1.3
    # via werkzeug
mysql-connector-python==8.0.33
    # via -r requirements.in
packaging==23.1
    # via
    #   build
    #   pytest
pip-tools==6.14.0
    # via -r requirements.in
pluggy==1.2.0
    # via pytest
protobuf==3.20.3
    # via mysql-connector-python
pyproject-hooks==1.0.0
    # via build
pytest==7.4.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
referencing==0.29.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.8.10
    # via
    #   jsonschema
    #   referencing
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pytest
toolz==0.12.0
    # via -r requirements.in
werkzeug==2.3.6
    # via -r requirements.in
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
